fracture: split payment sheet and projects view panels
